Cast tellg return value to int before doing operation
This is an attempt to solve the following error (caused on clang compiler): use of overloaded operator '-' is ambiguous (with operand types 'std::basic_istream<char, std::char_traits<char> >::pos_type' (aka 'fpos<__mbstate_t>') and 'const int')

+++ include/avl_database.hpp
@@ -353,7 +353,7 @@ class AvlDatabase
       tree_file.write(reinterpret_cast<char*>(node_ptr), sizeof(Node));
       tree_file.flush();
 
-      return (tree_file.tellg() - tree_file_offset) / sizeof(Node) - 1;
+      return ((int)(tree_file.tellg()) - tree_file_offset) / sizeof(Node) - 1;
     }
 
     /**
